+++ assets/scripts/core/module.d.ts
@@ -1,4 +1,4 @@
-/// <reference path='../../vendor/grottopress/jentil/assets/scripts/global.d.ts' />
+/// <reference path='../../../vendor/grottopress/jentil/assets/scripts/core/module.d.ts' />
 
 /*
 |--------------------------------------------------------------------------
@@ -6,15 +6,14 @@
 |--------------------------------------------------------------------------
 |
 | A declaration is how you tell the Typescript compiler that you are going
-| to use an object, method, variable etc. not defined/imported in the
-| current module/file.
+| to use an item not defined or imported in the current module.
 |
 | They are sort of like C/C++ header files, and consists only of member
 | signatures.
 |
-| See `assets/scripts/global.d.ts` in Jentil for examples.
+| See `assets/scripts/customize-preview/module.d.ts` in Jentil for examples.
 |
-| Add `/// <reference path="./global.d.ts" />` in any new `.ts` file, to use
+| Add `/// <reference path="./module.d.ts" />` in any new `.ts` file, to use
 | the declarations here in that file.
 |
 | Read more:
@@ -22,4 +21,4 @@
 |
 */
 
-declare const myThemeAwesomePostsHeadingModId: string
+// declare const myVar: string