+++ assets/scripts/customize-preview/module.d.ts
@@ -1,0 +1,3 @@
+/// <reference path='../../../vendor/grottopress/jentil/assets/scripts/customize-preview/module.d.ts' />
+
+declare const myThemeAwesomePostsHeadingModId: string