+++ gulpfile.js
@@ -1,88 +1,36 @@
 'use strict'
 
-const { src, dest, watch, series, parallel } = require('gulp')
+const { dest, parallel, series, src, watch } = require('gulp')
 
-const uglify = require('gulp-uglify')
-const rename = require('gulp-rename')
-const rtlcss = require('gulp-rtlcss')
-const sass = require('gulp-sass')
-const sourcemaps = require('gulp-sourcemaps')
-const typescript = require('gulp-typescript')
-const postcss = require('gulp-postcss')
 const cssnano = require('cssnano')
+const filter = require('gulp-filter')
 const focus = require('postcss-focus')
 const newer = require('gulp-newer')
+const postcss = require('gulp-postcss')
+const rename = require('gulp-rename')
+const rollup = require('gulp-better-rollup')
+const rtlcss = require('gulp-rtlcss')
+const sass = require('gulp-sass')
 const sh = require('shelljs')
+const uglify = require('gulp-uglify')
 
-const tsConfig = require('./tsconfig.json')
+const roConfig = require('./rollup.config')
+
+const uglifyOpts = {output: {comments: /(^!|\@license|\@preserve)/i}}
 
 const paths = {
-    styles: {
-        src: ['./assets/styles/**/*.scss'],
-        dest: './dist/styles'
-    },
     scripts: {
-        src: tsConfig.include,
-        dest: tsConfig.compilerOptions.outDir
+        dest: roConfig.output.dir,
+        mapDest: '.',
+        src: roConfig.input,
+        watchSrc: ['./assets/scripts/**/*.ts']
     },
-    vendor: {
-        dest: {
-            dist: './dist/vendor',
-            assets: './assets/vendor'
-        }
+    styles: {
+        dest: './dist/styles',
+        mapDest: '.',
+        src: ['./assets/styles/*.scss'],
+        watchSrc: ['./assets/styles/**/*.scss']
     }
-}
-
-function _scripts(done)
-{
-    src(paths.scripts.src)
-        .pipe(newer(paths.scripts.dest))
-        .pipe(sourcemaps.init())
-        .pipe(typescript(tsConfig.compilerOptions))
-        .pipe(uglify())
-        .pipe(rename({'suffix': '.min'}))
-        .pipe(sourcemaps.write())
-        .pipe(dest(paths.scripts.dest))
-
-    done()
-}
-
-function _styles(done)
-{
-    src(paths.styles.src)
-        .pipe(newer(paths.styles.dest))
-        .pipe(sourcemaps.init())
-        .pipe(sass().on('error', sass.logError))
-        .pipe(postcss([focus(), cssnano()]))
-        .pipe(rename({'suffix': '.min'}))
-        .pipe(sourcemaps.write())
-        .pipe(dest(paths.styles.dest))
-        .pipe(rtlcss())
-        .pipe(rename(path =>
-            path.basename = path.basename.replace('.min', '-rtl.min')
-        ))
-        .pipe(dest(paths.styles.dest))
-
-    done()
-}
-
-function _watch()
-{
-    watch(paths.scripts.src, {ignoreInitial: false}, _scripts)
-    watch(paths.styles.src, {ignoreInitial: false}, _styles)
-}
-
-function _clean(done)
-{
-    sh.rm(
-        '-rf',
-        paths.styles.dest,
-        paths.scripts.dest,
-        paths.vendor.dest.dist,
-        paths.vendor.dest.assets
-    )
-
-    done()
 }
 
 function _chmod(done)
@@ -92,10 +40,59 @@ function _chmod(done)
     done()
 }
 
-exports.styles = _styles
-exports.scripts = _scripts
-exports.watch = _watch
-exports.clean = _clean
+function _clean(done)
+{
+    sh.rm(
+        '-rf',
+        paths.styles.dest,
+        paths.scripts.dest
+    )
+
+    done()
+}
+
+function _scripts(done)
+{
+    src(paths.scripts.src, {sourcemaps: true})
+        .pipe(newer(paths.scripts.dest))
+        .pipe(rollup({plugins: roConfig.plugins}, roConfig.output))
+        .pipe(rename({'suffix': '.min', 'extname': '.js'}))
+        .pipe(uglify(uglifyOpts))
+        .pipe(dest(paths.scripts.dest, {sourcemaps: paths.scripts.mapDest}))
+
+    done()
+}
+
+function _styles(done)
+{
+    src(paths.styles.src, {sourcemaps: true})
+        .pipe(newer(paths.styles.dest))
+        .pipe(sass().on('error', sass.logError))
+        .pipe(postcss([focus(), cssnano()]))
+        .pipe(rename({'suffix': '.min'}))
+        .pipe(dest(paths.styles.dest, {sourcemaps: paths.styles.mapDest}))
+        .pipe(filter(['**/*.css']))
+        .pipe(rtlcss())
+        .pipe(rename(path => {
+            path.basename = path.basename.replace('.min', '-rtl.min')
+        }))
+        .pipe(dest(paths.styles.dest, {sourcemaps: paths.styles.mapDest}))
+
+    done()
+}
+
+function _watch(done)
+{
+    watch(paths.scripts.watchSrc, {ignoreInitial: false}, _scripts)
+    watch(paths.styles.watchSrc, {ignoreInitial: false}, _styles)
+
+    done()
+}
+
 exports.chmod = _chmod
+exports.clean = _clean
+exports.scripts = _scripts
+exports.styles = _styles
+exports.watch = _watch
 
 exports.default = series(parallel(_styles, _scripts), _watch)